+++ mitra_models/src/filter_rules/types.rs
@@ -12,6 +12,7 @@ pub enum FilterAction {
     RejectProfileImages,
     RejectCustomEmojis,
     RejectData,
+    MarkSensitive,
 }
 
 impl From<FilterAction> for i16 {
@@ -22,6 +23,7 @@ impl From<FilterAction> for i16 {
             FilterAction::RejectProfileImages => 3,
             FilterAction::RejectCustomEmojis => 4,
             FilterAction::RejectData => 5,
+            FilterAction::MarkSensitive => 6,
         }
     }
 }
@@ -36,6 +38,7 @@ impl TryFrom<i16> for FilterAction {
             3 => Self::RejectProfileImages,
             4 => Self::RejectCustomEmojis,
             5 => Self::RejectData,
+            6 => Self::MarkSensitive,
             _ => return Err(DatabaseTypeError),
         };
         Ok(action)